My Organization
-Mappers
-Client side model
-view model,data service,View added

+++ Interfaces/Repository/IBaseRepository.cs
@@ -52,13 +52,13 @@ namespace MPC.Interfaces.Repository
         /// <summary>
         /// User Domain Key
         /// </summary>
-        long UserDomainKey { get;}
+        int UserDomainKey { get; }
 
         /// <summary>
         /// Logged in user identity
         /// </summary>
         string LoggedInUserIdentity { get; }
 
-        
+
     }
 }
